i3 - fixed display problems related to resize functionality
