+++ src/graphiql.css
@@ -46,6 +46,8 @@
 .graphiql-container .topBarWrap {
   display: flex;
   flex-direction: row;
+  flex-wrap: nowrap;
+  min-width: 640px;
 }
 
 .graphiql-container .topBar {
