+++ src/graphiql.js
@@ -1,11 +1,10 @@
-import React, {useState, useRef} from 'react';
+import React, {useState, useEffect, useRef} from 'react';
 import GraphiQL from 'graphiql';
 import fetch from 'isomorphic-fetch';
-import MetaQueryInput from './MetaQueryInput'
+import GraphiQLMetaFilters from './GraphiQLMetaFilters'
 import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
 import Slide from '@material-ui/core/Slide';
-import Fade from '@material-ui/core/Fade';
 import './graphiql.css';
 import './cgraphiql.css'
 
@@ -13,14 +12,6 @@ const useStyles = makeStyles(theme => ({
   gridContainer: {
     width: "100%",
     height: `calc(100vh - 34px)`,
-  },
-  gridItemGraphiql: {
-    width: "100%",
-    flex: "1",
-  },
-  gridItemMetafilter: {
-    width: "100%",
-    flex: "0",
   },
 }));
 
@@ -30,16 +21,21 @@ const metaquery_url = process.env.REACT_APP_SERVER_METAQUERY_URL || 'http://loca
 export default function MyGraphiQL(props){
   const classes = useStyles();
   const { loginHandler } = props;
-  const [gridItemMetafilterHeight, setGridItemMetafilterHeight] = useState(0);
-  const [metaQueryResult, setMetaQueryResult] = useState("");
+
+  const [graphiQLflexGrow, setGraphiQLflexGrow] = useState(1);
+  const [filterElementHeight, setFilterElementHeight] = useState(0);
   const [selectedFilter, setSelectedFilter] = useState("");
   const [hasFilter, setHasFilter] = useState(false);
+
+  const graphiQLflexGrowRef = useRef(1);
+  const filterElementHeightRef = useRef(0);
   const selectedFilterRef = useRef("");
-  const runMetaQueryRef = useRef(false);
-  const metaQueryFilterRef = useRef("");
+  const filterValueRef = useRef("");
+  const updateFlexLockRef = useRef(false);
+
   const graphiQL = useRef(null);
-  const gridItemMetafilterRef = useRef(null);
-  const gridItemGraphiqlRef = useRef(null);
+  const filterElementRef = useRef(null);
+  const graphiqlElementRef = useRef(null);
 
   const checkLoggin = () => {
     let expires = new Date(localStorage.getItem('expirationDate')) < new Date();
@@ -79,8 +75,8 @@ export default function MyGraphiQL(props){
     //set metaQuery parameters
     let metaQueryParams = {
       queries:   graphQLParams,
-      jq:        selectedFilterRef.current==='jq' ? metaQueryFilterRef.current: null,
-      jsonPath:  selectedFilterRef.current==='JsonPath' ? metaQueryFilterRef.current: null,
+      jq:        selectedFilterRef.current==='jq' ? filterValueRef.current: null,
+      jsonPath:  selectedFilterRef.current==='JsonPath' ? filterValueRef.current: null,
     };
     let headers = getHeaders();
 
@@ -93,6 +89,25 @@ export default function MyGraphiQL(props){
     });
   };
 
+  /**
+   * Effects
+   */
+  useEffect(() => {
+    //define resize handler
+    let handleWindowsResize = function () {
+      setFilterElementHeight(filterElementRef.current.clientHeight);
+    };
+    //set filter height
+    if(filterElementRef&&filterElementRef.current&&hasFilter) {
+      setFilterElementHeight(filterElementRef.current.clientHeight);
+      //add event listener
+      window.addEventListener("resize", handleWindowsResize);
+    } else {
+      //remove event listener
+      window.removeEventListener("resize", handleWindowsResize);
+    }
+  }, [hasFilter]);
+  
   /**
    * Handlers
    */
@@ -122,7 +137,7 @@ export default function MyGraphiQL(props){
 
   const handleRunMetaQuery = async (filter) => {
     //set meta-filter
-    metaQueryFilterRef.current = filter ? filter : null;
+    filterValueRef.current = filter ? filter : null;
 
     if(graphiQL.current) {
       //graphql params
@@ -163,19 +178,88 @@ export default function MyGraphiQL(props){
     selectedFilterRef.current = ("");
   }
 
-  const handleResize = () => {
-    if(gridItemMetafilterRef.current) {
-      setGridItemMetafilterHeight(gridItemMetafilterRef.current.clientHeight+1);
+  const onInitVerticalResize = (mouseDownEvent) => {
+    //checks
+    if(!mouseDownEvent || typeof mouseDownEvent !== 'object') return;
+    if(!graphiqlElementRef || !graphiqlElementRef.current) return;
+
+    //set initial values
+    let initialY = mouseDownEvent.clientY;
+    let initialHeight = graphiqlElementRef.current.clientHeight;
+
+    let handleMouseMove = function(mouseMoveEvent) {
+      //check: no left-button down
+      if(!mouseMoveEvent.buttons) {
+        document.removeEventListener("mousemove", handleMouseMove, true);
+        return;    
+      }
+      //set max height to current viewport height - topBar.
+      let maxHeight = (Math.max(document.documentElement.clientHeight || 0, window.innerHeight || 0) - 34);
+      //new values
+      let newHeight = initialHeight - (initialY - mouseMoveEvent.clientY); 
+      let newFilterHeight = maxHeight - newHeight;
+      let newFlexGrow = newHeight / newFilterHeight;
+      //check limits
+      if(newHeight > 100 && newHeight <= (maxHeight-34)) {
+        //update flex-grow
+        graphiQLflexGrowRef.current = newFlexGrow;
+        //update filter height
+        filterElementHeightRef.current = newFilterHeight;
+        //delayed state update (debounce)
+        if(updateFlexLockRef.current === false) {
+          updateFlexLockRef.current = true;
+          delayedUpdateFlex(70);
+        }
+      }
+    }
+    let handleMouseUp = function () {
+      document.removeEventListener("mousemove", handleMouseMove, true);
+      document.removeEventListener('mouseup', handleMouseUp);
+    };
+
+    if(mouseDownEvent.button === 0) { //left button
+      mouseDownEvent.preventDefault();
+      document.addEventListener("mousemove", handleMouseMove, true);
+      document.addEventListener('mouseup', handleMouseUp);
+    } else {
+      document.removeEventListener("mousemove", handleMouseMove, true);
+      document.removeEventListener('mouseup', handleMouseUp);
     }
   }
+
+  /**
+   * Utils
+   */
+  const delayedUpdateFlex = async (ms) => {
+    await new Promise(resolve => {
+      //set timeout
+      window.setTimeout(function() {
+        updateFlexLockRef.current = false;
+        setFilterElementHeight(filterElementHeightRef.current);
+        setGraphiQLflexGrow(graphiQLflexGrowRef.current);
+        resolve("ok");
+      }, ms);
+    });
+  };
 
   return (
     <div>
       <Grid container className={classes.gridContainer} spacing={0} direction="column" >
-        <Grid item 
-          style={hasFilter ? { maxHeight: `calc(100% - ${gridItemMetafilterHeight}px)` } : {maxHeight: '100%'}} 
-          className={classes.gridItemGraphiql}
-          ref={gridItemGraphiqlRef}
+        <div ref={graphiqlElementRef}
+          style={{
+            height: hasFilter ? "100%" : `calc(100vh - 34px)`,
+            width: "100%",
+            flex: graphiQLflexGrow,
+            WebkitFlex: graphiQLflexGrow,
+            transition: "flex .01s",
+            WebkitTransition: "flex .01s",
+            MozTransition: "flex .01s",
+            OTransition: "flex .01s",
+            transition: "height .05s",
+            WebkitTransition: "height .05s",
+            MozTransition: "height .05s",
+            OTransition: "height .05s",
+          }}
         >
           <GraphiQL
             ref={graphiQL}
@@ -209,20 +293,32 @@ export default function MyGraphiQL(props){
               />
             </GraphiQL.Toolbar>
           </GraphiQL>
-        </Grid>
-        <Grid item  className={classes.gridItemMetafilter} ref={gridItemMetafilterRef}>
-            <Slide direction="up" in={hasFilter} mountOnEnter unmountOnExit >
-              <div>
-                <MetaQueryInput 
-                  selectedFilter={selectedFilter}
-                  metaQueryResult={metaQueryResult}
-                  handleResize={handleResize}
-                  handleFilterSelected={handleFilterSelected}
-                  handleRunMetaQuery={handleRunMetaQuery}
-                  handleCloseFilter={handleCloseFilter} />
-              </div>
-            </Slide>
-        </Grid>
+        </div>
+        
+        <div ref={filterElementRef}
+          style={{
+            minHeight: 1,
+            width: "100%",
+            flex: "1 1",
+            WebkitFlex: "1, 1",
+            transition: "flex .01s",
+            WebkitTransition: "flex .01s",
+            MozTransition: "flex .01s",
+            OTransition: "flex .01s",
+          }}
+        >
+          <Slide direction="up" in={hasFilter} mountOnEnter unmountOnExit >
+            <div>
+              <GraphiQLMetaFilters 
+                selectedFilter={selectedFilter}
+                filterHeight={filterElementHeight}
+                onInitVerticalResize={onInitVerticalResize}
+                handleFilterSelected={handleFilterSelected}
+                handleRunMetaQuery={handleRunMetaQuery}
+                handleCloseFilter={handleCloseFilter} />
+            </div>
+          </Slide>
+        </div>
       </Grid>
     </div>
   );
